write

+++ c-skeleton/headers/write_csv.h
@@ -15,7 +15,7 @@
  * @param k Nombre total de centroïdes à écrire.
  * @param dimension Dimension des points dans les centroïdes.
  */
-void write_centroid(FILE *file, point_t* centroid, int64_t k, int64_t dimension);
+void write_centroid(FILE *file, point_t* centroid, int32_t k, uint32_t dimension);
 
 /**
  * Écrit les clusters dans un fichier au format spécifié.
@@ -25,7 +25,7 @@ void write_centroid(FILE *file, point_t* centroid, int64_t k, int64_t dimension)
  * @param k Nombre total de clusters.
  * @param dimension Dimension des points dans les clusters.
  */
-void write_cluster(FILE *file, cluster_t **cluster, int64_t k, int64_t dimension); 
+void write_cluster(FILE *file, cluster_t **cluster, int32_t k, uint32_t dimension); 
 
 /**
  * @brief Écrit les données de clustering dans un fichier CSV.
@@ -39,5 +39,5 @@ void write_cluster(FILE *file, cluster_t **cluster, int64_t k, int64_t dimension
  * @param dimension Nombre de dimensions des données
  * @param nombre_comb Nombre total de combinaisons
  */
-void write_csv(FILE *output_file, uint64_t* distortion, point_t **centroid_init_Array, point_t **centroid_final_Array, cluster_t ***clustersArray, int64_t k, int64_t dimension, int64_t nombre_comb, bool quiet_mode ); 
+void write_csv(FILE *output_file, uint64_t* distortion, point_t **centroid_init_Array, point_t **centroid_final_Array, cluster_t ***clustersArray, int32_t k, uint32_t dimension, int64_t nombre_comb, bool quiet_mode ); 
 #endif //WRITE_CSV_H
